move routing logic to routes directory
